requirements.txt
opentelemetry-api==1.5.0
opentelemetry-exporter-zipkin==1.5.0
opentelemetry-exporter-zipkin-json==1.5.0
opentelemetry-exporter-zipkin-proto-http==1.5.0
opentelemetry-exporter-jaeger>=0.1.0
opentelemetry-instrumentation==0.24b0
opentelemetry-sdk==1.5.0
opentelemetry-semantic-conventions==0.24b0
protobuf==3.17.3
requests==2.26.0
six==1.16.0
urllib3==1.26.6
wrapt==1.12.1
